fix datastore migration error

+++ eloq_data_store_service/data_store_service_config.cpp
@@ -865,7 +865,7 @@ bool DataStoreServiceClusterManager::SwitchShardToReadWrite(
 
     topology_.UpdateDSShardStatus(shard_id, DSShardStatus::ReadWrite);
     DLOG(INFO) << "SwitchToReadWrite, shard " << shard_id
-               << " status: " << shard_status;
+               << " status: " << static_cast<int>(shard_status);
     return true;
 }
